Move client modules back into Meteor directory.
